% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.mclist}
\alias{as.mclist}
\alias{as.mclist.glm}
\alias{as.mclist.glmerMod}
\alias{as.mclist.list}
\alias{as.mclist.lm}
\alias{as.mclist.mcmc.list}
\alias{as.mclist.merMod}
\alias{as.mclist.polr}
\alias{as.mclist.sim}
\alias{as.mclist.stanfit}
\alias{is.mclist}
\title{mclist coercion}
\usage{
as.mclist(x, ...)

\method{as.mclist}{stanfit}(x, ...)

\method{as.mclist}{sim}(sims, ...)

\method{as.mclist}{lm}(x, n_sims = 100)

\method{as.mclist}{glm}(x, n_sims = 100)

\method{as.mclist}{polr}(x, n_sims = 100)

\method{as.mclist}{merMod}(x, n_sims = 100)

\method{as.mclist}{glmerMod}(x, n_sims = 100)

\method{as.mclist}{list}(x, ...)

\method{as.mclist}{mcmc.list}(x, ...)

is.mclist(x)
}
\arguments{
\item{n_sims}{Number of simulations to obtain from a fitted model.}
}
\value{
An mclist. This is a list of posterior simulations with one
element per parameter. Each list element is an array whose the
first dimension represents the simulations.
}
\description{
Converts posterior simulations and fitted models to an
mclist. Simulations from fitted models are obtained through
\code{sim} from the package arm.
}
\section{Methods (by class)}{
\itemize{
\item \code{stanfit}: Converts Stan simulations to mclist

\item \code{sim}: Converts arm::sim simulations to mclist

\item \code{lm}: Simulate from a lm fitted model and convert
to mclist

\item \code{glm}: Simulates from a glm fitted model and
converts to mclist

\item \code{polr}: Simulates from a glm fitted model and
converts to mclist

\item \code{merMod}: Simulates from a lmer fitted model and
converts to mclist

\item \code{glmerMod}: Simulates from a glmer fitted model and
converts to mclist

\item \code{list}: Converts Jags simulations to mclist

\item \code{mcmc.list}: Converts Coda simulations to mclist
}}

