% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/verbs-storage.R
\name{as_matrix}
\alias{as_matrix}
\title{Assemble a list of vectors and matrices into a matrix}
\usage{
as_matrix(...)
}
\arguments{
\item{...}{Vectors and/or lists of vectors.}
}
\description{
\code{as_matrix()} binds columns into a matrix. These columns can
come in the form of vectors or of a list or a data frame of
vectors.
}
\examples{
vectors <- mtcars[3:4]
as_matrix(vectors, mpg = mtcars$mpg, vectors, mtcars$am)

# If applied on an empty list or without arguments, as_matrix()
# returns a 0x0 matrix:
matrix()
}

