% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools-modeling.R
\name{Bernoulli}
\alias{Bernoulli}
\alias{dbernoulli}
\alias{pbernoulli}
\alias{qbernoulli}
\alias{rbernoulli}
\title{Bernoulli distribution wrappers}
\usage{
dbernoulli(x, prob, log = FALSE)

pbernoulli(q, prob, lower.tail = TRUE, log.p = FALSE)

qbernoulli(p, prob, lower.tail = TRUE, log.p = FALSE)

rbernoulli(n, prob)
}
\arguments{
\item{x}{vector of quantiles.}

\item{prob}{probability of success on each trial.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are
    \eqn{P[X \le x]}, otherwise, \eqn{P[X > x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1}, the length
    is taken to be the number required.}
}
\description{
Wrappers for dbinom, pbinom, qbinom and rbinom with \code{size} set
to 1.
}

