% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functional-by-sim.R
\name{by_sim}
\alias{by_sim}
\alias{by_sim.list}
\alias{by_sim.sims}
\alias{by_sim.sims_array}
\title{Map a function simulation-wise}
\usage{
by_sim(.x, .f, ..., .context = list(), .bindings = NULL)

\method{by_sim}{list}(.x, .f, ..., .context = list(), .bindings = NULL)

\method{by_sim}{sims}(.x, .f, ..., .context = list(), .bindings = NULL)

\method{by_sim}{sims_array}(.x, .f, ..., .context = list(), .sims = list(),
  .bindings = NULL)
}
\arguments{
\item{.x}{A list of simulations arrays, or a simulations array.}

\item{.f}{A function or an expression to be applied to each
simulation.}

\item{...}{Further arguments passed on to \code{.f}.}

\item{.context}{A list or a data frame that contains additional
variables used in \code{.f}.}

\item{.bindings}{A named vector linking the arguments of \code{.f}
to the names of the simulations arrays in \code{.x}, in case they
do not match.}

\item{.sims}{A list of simulations arrays containing additional
parameters on which \code{.statistic} depends.}
}
\description{
\code{by_sim()} is similar to \code{apply()} in that it applies a
function \code{.f} to the margins of an array. \code{by_sim()} is
specialised to apply the function on each simulation of a
\code{sims_array}. It can be also be applied to a list of
simulations, in which case \code{.f} may depend on several
parameters.
}

