% Generated by roxygen2 (4.0.2): do not edit by hand
\name{col_vector}
\alias{col_vector}
\title{Make column vector}
\usage{
col_vector(x)
}
\arguments{
\item{x}{numeric vector.}
}
\value{
Array of dimension n x 1.
}
\description{
Wrapper around cbind to produce a column vector. \code{col_vector(x)}
is equivalent to \code{cbind(x)} but it only accepts one argument
and is a clearer semantic signal.
}

