% Generated by roxygen2 (4.0.2): do not edit by hand
\name{colVars}
\alias{colVars}
\title{Optimised variance computation}
\usage{
colVars(x)
}
\arguments{
\item{x}{numeric matrix.}
}
\description{
colVars uses colMeans to provide an efficient way of computing the
variances of the columns of a matrix.
}

