% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools-arrays.R
\name{colwise-rowise}
\alias{cols_means}
\alias{cols_variances}
\alias{colwise-rowise}
\alias{rows_means}
\alias{rows_variances}
\title{Colwise and rowwise operations}
\usage{
cols_variances(x)

rows_variances(x)

cols_means(x, na.rm = FALSE, dims = 1L)

rows_means(x, na.rm = FALSE, dims = 1L)
}
\arguments{
\item{x}{numeric matrix.}
}
\description{
\code{cols_means()} and \code{rows_means()} are simple aliases to
\code{colMeans()} and \code{rowMeans()}. \code{colVars()} uses
\code{colMeans()} to provide an efficient way of computing the
variances of the columns of a matrix while \code{rowVars()} uses
\code{rowMeans()} to provide an efficient way of computing the
variances of the rows of a matrix.
}

