% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/verbs-storage.R
\name{define}
\alias{define}
\alias{define_}
\alias{define_.data.frame}
\alias{define_.list}
\title{Define or redefine elements of a list}
\usage{
define(.data, ...)

define_(.data, ..., .dots)

\method{define_}{list}(.data, ..., .dots)

\method{define_}{data.frame}(.data, ..., .dots)
}
\arguments{
\item{.data}{A list or data frame.}

\item{...}{Variables to keep and new definitions.}
}
\description{
This function takes a list of objects or a data frame and changes
their contents. It discards all elements that are not mentioned in
\code{...}, leaving a clean list of inputs for your models.
}

