% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/verbs-storage.R
\name{design}
\alias{design}
\title{Assemble variables into a (design) matrix}
\usage{
design(..., .unjoin = NULL)
}
\arguments{
\item{...}{Named arguments.}

\item{.unjoin}{When a vector identifying groups, and if the
variables in \code{...} are unique within groups, the design
matrix is reduced to one value per group.}
}
\description{
\code{design()} works like dplyr's function
\code{data_frame()}. New columns can be defined in terms of
previous columns.
}
\details{
It converts factor variables to columns of indicators. It supports
the \code{:} operator to interact predictors. Only variables of
length 1 are recycled.
}

