% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/examples.R
\name{examples}
\alias{examples}
\alias{radon_data}
\alias{radon_model}
\alias{wells_data}
\title{Example datasets and models}
\usage{
wells_data()

radon_data()

radon_model(language = "stan", variant = "centered", ref = FALSE)
}
\description{
Obtain and tidy ARM's wells and radon datasets from
http://github.com/stan-dev/example-models/
}

