% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fetch}
\alias{fetch}
\alias{fetch_radon_data}
\alias{fetch_wells_data}
\title{Fetch ARM datasets}
\usage{
fetch_wells_data()

fetch_radon_data()
}
\description{
Obtain and tidy ARM's wells and radon datasets from
http://github.com/stan-dev/example-models/
}

