% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{gs_chr_names}
\alias{gs_chr_names}
\title{Creates names}
\usage{
gs_chr_names(.data, colname)
}
\arguments{
\item{.data}{(required) the source data}

\item{colname(required)}{the column name for this created col}
}
\value{
appended dataframe with a char  vector added as the last col
}
\description{
Creates names
}
