% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{gs_fct}
\alias{gs_fct}
\title{Splits a set of factors as per the splits provided}
\usage{
gs_fct(.data, colname, fcts, splits)
}
\arguments{
\item{.data}{(required) the source data}

\item{colname}{(required) the column name for this created col}

\item{fcts}{(required) the factors that need to be expanded into rows based on splits}

\item{splits}{(required) the % splits for factors. should add to 1.0}
}
\value{
the appended dataframe with a numeric vector added as the last col
}
\description{
Proving fcts = c("M" , "F") with splits = c(.3,.7) will randomly allocate 30% rows as Males and 70% as femails
}
\examples{
gs_fct(.data = mtcars , colname = "sex", fcts = c("M", "F"), splits = c(.3, .7))
}
