% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{gs_id_seq}
\alias{gs_id_seq}
\title{Creates sequencial ids with starting point at start and n = nrows(.data)}
\usage{
gs_id_seq(.data, colname, start = 1)
}
\arguments{
\item{.data}{(required) the source data}

\item{start}{(optional) default = 1}

\item{colname(required)}{the column name for this created col}
}
\value{
appended dataframe with a numeric vector added as the last col
}
\description{
Creates sequencial ids with starting point at start and n = nrows(.data)
}
\examples{
gs_id_seq(data, fcts = c("M", "F"), splits = c(.5, .5), colname = "sex")
}
