% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{gs_norm}
\alias{gs_norm}
\title{Returns a normally distributed vector with mean and SD , with min and max}
\usage{
gs_norm(.data, colname, mu, sd, round_to_digits = 0, min = NULL, max = NULL)
}
\arguments{
\item{.data}{(required) the source data}

\item{colname}{(optional) the col name for vector added to .data. Recommeded to set else will set at num1}

\item{mu}{(required) the mean}

\item{sd}{(required) standard deviation}

\item{round_to_digits}{(optional) default is 0.}

\item{min}{(optional) min cut of}

\item{maximum}{(optional) max vut of}
}
\value{
appended dataframe with a numeric vector added as the last col
}
\description{
Returns a normally distributed vector with mean and SD , with min and max
}
\examples{
 gs_norm(data , colname = "age" , mu = 45 , sd = 20 , min = 18 , max =65 )
}
