% Generated by roxygen2 (4.0.2): do not edit by hand
\name{intercept}
\alias{intercept}
\title{Intercept column}
\usage{
intercept(n = NULL)
}
\arguments{
\item{n}{}
}
\description{
Create a column of ones. This is useful to make a design matrix.
}
\details{
\code{n} is an optional argument: when \code{intercept} is nested
inside \code{cbind}, \code{data.frame} or \code{data_frame}, it
examines its surroundings to automatically determine its length.
}
\examples{
cbind(intercept(), 1:5)
}

