% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/verbs-mutate.R
\name{mutate_.sims}
\alias{mutate_.sims}
\title{Transform simulations arrays}
\usage{
\method{mutate_}{sims}(.data, ..., .dots, .context = list())
}
\arguments{
\item{.data}{A list of simulations arrays.}

\item{...}{Name-value pairs of expressions.}

\item{.dots}{Used to work around non-standard evaluation. See
dplyr's \code{vignette("nse")} for details.}

\item{.context}{Additional data on which expressions depend.}
}
\description{
Performs computations datawise.
}
\seealso{
dplyr's \code{\link[dplyr]{mutate}()}
}

