% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sims.R
\name{permute_sims}
\alias{permute_sims}
\alias{permute_sims.sims}
\alias{permute_sims.sims_array}
\title{Rearrange the dimensions of an MCMC array}
\usage{
permute_sims(sims, to_major = TRUE)

\method{permute_sims}{sims}(sims, to_major = TRUE)

\method{permute_sims}{sims_array}(x, to_major = TRUE)
}
\arguments{
\item{sims}{Either a list of simulation arrays, or one such array.}

\item{to_major}{If \code{TRUE}, the array is permuted so that the
first dimension becomes the last one. \code{FALSE} yields the
opposite permutation.}
}
\description{
\code{permute_sims()} rearranges the dimensions of an array,
switching between simulations-major order (where simulations are
indexed by the first array dimension) and data-major order (where
they are indexed by the last dimension). If \code{to_major} is
\code{NULL}, the \code{sims_major_order} attribute is consulted to
determine in which dimension the simulations are currently stored.
}
\details{
Indexing the simulations in the right dimension is important when
performance is critical. Data-wise operations (such as taking a
linear combination of regression coefficients ) are faster when the
first dimension indexes simulations. On the other hand,
simulation-wise operations (such as taking the posterior mean of a
quantity) are faster when the simulations are indexed along the
last dimension.
}
\section{Methods (by class)}{
\itemize{
\item \code{sims}: 

\item \code{sims_array}: 
}}

