% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsim.R
\name{print.gsim}
\alias{print.gsim}
\title{Print a gsim object}
\usage{
\method{print}{gsim}(x, ...)
}
\arguments{
\item{x}{A \code{gsim} object.}

\item{...}{Unused.}
}
\value{
\code{print.gsim()} returns \code{x} invisibly.
}
\description{
Print a gsim object
}
