% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsim.R
\name{reduce_dataset}
\alias{reduce_dataset}
\title{reduces size of dataset provided by random subsetting}
\usage{
reduce_dataset(.data, target_rows)
}
\arguments{
\item{.data}{input data}

\item{target_rows}{the reduced nrow(data) needed}
}
\description{
reduces size of dataset provided by random subsetting
}
