% Generated by roxygen2 (4.0.2): do not edit by hand
\name{row_vector}
\alias{row_vector}
\title{Make row vector}
\usage{
row_vector(x)
}
\arguments{
\item{x}{numeric vector.}
}
\value{
Array of dimension 1 x n.
}
\description{
Wrapper around rbind to produce a row vector. \code{row_vector(x)}
is equivalent to \code{rbind(x)} but it only accepts one argument
and is a clearer semantic signal.
}

