% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rowVars}
\alias{rowVars}
\title{Optimised variance computation}
\usage{
rowVars(x)
}
\arguments{
\item{x}{a numeric matrix}
}
\description{
rowVars uses rowMeans to provide an efficient way of computing the
variances of the rows of a matrix.
}

