% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/verbs-select.R
\name{select_.sims}
\alias{rename_.sims}
\alias{rename_.stanfit}
\alias{select_.sims}
\alias{select_.stanfit}
\title{Select/rename parameters}
\usage{
\method{select_}{sims}(.data, ..., .dots)

\method{select_}{stanfit}(.data, ..., .dots)

\method{rename_}{sims}(.data, ..., .dots)

\method{rename_}{stanfit}(.data, ..., .dots)
}
\arguments{
\item{.data}{A list of simulations arrays.}

\item{...}{Name-value pairs of expressions.}

\item{.dots}{Used to work around non-standard evaluation. See
dplyr's \code{vignette("nse")} for details.}
}
\description{
Select parameters from a list of
simulations. \code{\link{select}()} keeps only the variables
mentioned in \code{...}. \code{\link{rename}()} keeps all
variables.
}
\section{Special functions}{

From dplyr's man page: As well as using existing functions like
\code{:} and \code{c}, there are a number of special functions that
only work inside \code{select}

\itemize{
 \item \code{starts_with(x, ignore.case = TRUE)}:
   names starts with \code{x}
 \item \code{ends_with(x, ignore.case = TRUE)}:
   names ends in \code{x}
 \item \code{contains(x, ignore.case = TRUE)}:
   selects all variables whose name contains \code{x}
 \item \code{matches(x, ignore.case = TRUE)}:
   selects all variables whose name matches the regular expression \code{x}
 \item \code{num_range("x", 1:5, width = 2)}:
   selects all variables (numerically) from x01 to x05.
 \item \code{one_of("x", "y", "z")}:
   selects variables provided in a character vector.
 \item \code{everything()}:
   selects all variables.
}
}
\seealso{
dplyr's \code{\link[dplyr]{select}()} and
  \code{\link[dplyr]{rename}()}
}

