% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/verbs-mutate.R
\name{sims-verbs}
\alias{mutate_.stanfit}
\alias{sims-verbs}
\title{Verbs for manipulating posterior simulations}
\usage{
\method{mutate_}{stanfit}(.data, ..., .dots, .context = list())
}
\arguments{
\item{.data}{A list of simulations arrays.}

\item{...}{Name-value pairs of expressions.}

\item{.dots}{Used to work around non-standard evaluation. See
dplyr's \code{vignette("nse")} for details.}

\item{.context}{Additional data on which expressions depend.}
}
\value{
A list of simulations arrays.
}
\description{
These functions provide dplyr-like methods for manipulating lists
of posterior simulations (\code{sims} object). See
\code{\link{mutate}()}, \code{\link{summarise}()} and
\code{\link{select}()} .
}

