% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sims.R
\name{sims_array}
\alias{sims_array}
\title{Construct a simulations array object}
\usage{
sims_array(array, major_order)
}
\arguments{
\item{array}{An array of simulations.}

\item{major_order}{If \code{TRUE}, this indicates that the array is
stored in simulations-major order as opposed to data-major order.}
}
\description{
Construct a simulations array object
}
\seealso{
\code{\link{permute_sims}()}
}

