% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools-arrays.R
\name{stan-compat}
\alias{quad_form_diag}
\alias{stan-compat}
\title{Stan compatibility}
\usage{
quad_form_diag(matrix, vector)
}
\arguments{
\item{matrix}{A matrix.}

\item{vector}{A vector.}
}
\description{
Stan functions that do not exist in R.
}

