% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/verbs-storage.R
\name{store}
\alias{store}
\alias{store_}
\title{Create a list}
\usage{
store(...)

store_(dots, data = list())
}
\arguments{
\item{List}{elements.}
}
\description{
This function is analog to \code{\link[dplyr]{data_frame}()} from
the dplyr package but creates a list instead of a data frame. It
evaluates its elements lazily and sequentially.
}

