% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools-modeling.R
\name{summaries}
\alias{mean_sd}
\alias{q025}
\alias{q975}
\alias{summaries}
\title{Compute summaries a quantity}
\usage{
mean_sd(x, rm_na = TRUE)

q025(x)

q975(x)
}
\arguments{
\item{x}{A quantity to summarise.}

\item{rm_na}{Whether to remove missing values.}
}
\description{
These tools are especially useful in conjunction with
\code{\link{summarise}()}. \code{mean_sd()} returns a list with
components \code{mean} and \code{sd}.
}
\examples{
radon_sims \%>\%
  summarise(
    mu_y = list(q025(mu_y), q975(mu_y)),
    Beta = mean_sd(Beta)
  )
}

