% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sims-arrays.R, R/sims.R
\name{to_sims_major}
\alias{is_sims_major}
\alias{to_data_major}
\alias{to_data_major.sims}
\alias{to_sims_major}
\alias{to_sims_major.sims}
\alias{to_sims_major.sims_array}
\title{Rearrange the dimensions of an MCMC array}
\usage{
to_sims_major(x)

\method{to_sims_major}{sims_array}(x)

to_data_major(x)

is_sims_major(x)

\method{to_sims_major}{sims}(x)

to_data_major.sims(x)
}
\arguments{
\item{x}{A sims or sims_array object.}
}
\description{
\code{to_data_major()} and \code{to_sims_major()} rearrange the
dimensions of an array so that the simulations are indexed along
the first dimension (data-major order) or the last
(simulation-major order).
}
\details{
Indexing the simulations in the right dimension is important when
performance is critical. Data-wise operations (such as taking a
linear combination of regression coefficients ) are faster when the
first dimension indexes simulations. On the other hand,
simulation-wise operations (such as taking the posterior mean of a
quantity) are faster when the simulations are indexed along the
last dimension.
}

