% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools-arrays.R
\name{vector}
\alias{col_vector}
\alias{row_vector}
\alias{vector}
\title{Make row vector}
\usage{
row_vector(x)

col_vector(x)
}
\arguments{
\item{x}{numeric vector.}
}
\value{
Array of dimension 1 x n or n x 1
}
\description{
Wrappers around \code{rbind()} and \code{cbind()} to produce a row
or column vector. They accept only one argument and are a clearer
semantic signal regarding the nature of the vector.
}

