% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.mclist}
\alias{as.mclist}
\alias{as.mclist.glm}
\alias{as.mclist.glmerMod}
\alias{as.mclist.list}
\alias{as.mclist.lm}
\alias{as.mclist.mcmc.list}
\alias{as.mclist.merMod}
\alias{as.mclist.polr}
\alias{as.mclist.sim}
\alias{as.mclist.stanfit}
\alias{is.mclist}
\title{mclist coercion}
\usage{
as.mclist(x, ...)

\method{as.mclist}{stanfit}(x, ...)

\method{as.mclist}{sim}(sims, ...)

\method{as.mclist}{lm}(x, n_sims = 100)

\method{as.mclist}{glm}(x, n_sims = 100)

\method{as.mclist}{polr}(x, n_sims = 100)

\method{as.mclist}{merMod}(x, n_sims = 100)

\method{as.mclist}{glmerMod}(x, n_sims = 100)

\method{as.mclist}{list}(x, ...)

\method{as.mclist}{mcmc.list}(x, ...)

is.mclist(x)
}
\arguments{
\item{n_sims}{Passed to sim from the arm package. Number of
simulations to obtain from a fitted model.}
}
\value{
An mclist. This is a list of posterior simulations with one
element per parameter. Each list element is an array whose the
first dimension represents the simulations.
}
\description{
Converts posterior simulations and fitted models to an
mclist. Simulations from fitted models are obtained through
\code{sim} from the package arm.
}
\section{Methods (by class)}{
\itemize{
\item \code{stanfit}: Convert Stan simulations to mclist

\item \code{sim}: Convert arm::sim simulations to mclist

\item \code{lm}: Simulate from a lm fitted model and convert
to mclist

\item \code{glm}: Simulate from a glm fitted model and
convert to mclist

\item \code{polr}: Simulate from a glm fitted model and
convert to mclist

\item \code{merMod}: Simulate from a lmer fitted model and
convert to mclist

\item \code{glmerMod}: Simulate from a glmer fitted model and
convert to mclist

\item \code{list}: Convert Jags or Stan simulation lists to mclist

\item \code{mcmc.list}: Convert Coda simulations to mclist
}}

