% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as-sims-bayes.R, R/sims.R
\name{as_sims.stanfit}
\alias{as_sims}
\alias{as_sims.jagslist}
\alias{as_sims.list}
\alias{as_sims.mcmc.list}
\alias{as_sims.sims}
\alias{as_sims.stanfit}
\alias{is_sims}
\alias{n_sims.sims}
\title{Coerce to a sims object}
\usage{
\method{as_sims}{stanfit}(x, sims_major = NULL, vars = NULL, ...)

\method{as_sims}{jagslist}(x, ...)

\method{as_sims}{mcmc.list}(x, ...)

as_sims(x, ...)

\method{as_sims}{sims}(x, ...)

\method{as_sims}{list}(x, sims_major, ...)

is_sims(x)

\method{n_sims}{sims}(x)
}
\arguments{
\item{x}{An object to convert.}

\item{sims_major}{When converting a list of arrays to a \code{sims}
object, this indicates whether your arrays are stored in
simulation-major order.}

\item{...}{Further arguments passed to coercion functions.}
}
\description{
Converts posterior simulations and fitted models to a \code{sims}
object, a list of simulation arrays.
}

