% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sims.R
\name{as_sims}
\alias{as_sims}
\alias{as_sims.glm}
\alias{as_sims.glmerMod}
\alias{as_sims.list}
\alias{as_sims.lm}
\alias{as_sims.mcmc.list}
\alias{as_sims.merMod}
\alias{as_sims.polr}
\alias{as_sims.sim}
\alias{as_sims.stanfit}
\alias{is.mclist}
\title{Coerce a set of MCMC arrays to a sims object}
\usage{
as_sims(x, major_order = NULL, ...)

\method{as_sims}{stanfit}(x, major_order = NULL, ...)

\method{as_sims}{sim}(x, major_order = NULL, ...)

\method{as_sims}{lm}(x, major_order = NULL, n_sims = 100)

\method{as_sims}{glm}(x, major_order = NULL, n_sims = 100)

\method{as_sims}{polr}(x, major_order = NULL, n_sims = 100)

\method{as_sims}{merMod}(x, major_order = NULL, n_sims = 100)

\method{as_sims}{glmerMod}(x, major_order = NULL, n_sims = 100)

\method{as_sims}{list}(x, major_order = NULL, ...)

\method{as_sims}{mcmc.list}(x, major_order = NULL, ...)

is.mclist(x)
}
\arguments{
\item{major_order}{If \code{TRUE}, the simulations array is
arranged in simulation-major order. If \code{FALSE}, the array is
arranged in data-major order. If \code{NULL}, the order of the
original data structure is retained.}

\item{n_sims}{Passed to sim from the arm package. Number of
simulations to obtain from a fitted model.}
}
\value{
An mclist. This is a list of posterior simulations with one
  element per parameter. Each list element is an array whose the
  first dimension represents the simulations.
}
\description{
Converts posterior simulations and fitted models to an
mclist. Simulations from fitted models are obtained through
\code{sim} from the package arm.
}
\section{Methods (by class)}{
\itemize{
\item \code{stanfit}: Convert Stan simulations to mclist

\item \code{sim}: Convert arm::sim simulations to mclist

\item \code{lm}: Simulate from a lm fitted model and convert
to mclist

\item \code{glm}: Simulate from a glm fitted model and
convert to mclist

\item \code{polr}: Simulate from a glm fitted model and
convert to mclist

\item \code{merMod}: Simulate from a lmer fitted model and
convert to mclist

\item \code{glmerMod}: Simulate from a glmer fitted model and
convert to mclist

\item \code{list}: Convert Jags or Stan simulation lists to mclist

\item \code{mcmc.list}: Convert Coda simulations to mclist
}}

