% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Bernoulli}
\alias{Bernoulli}
\alias{dbernoulli}
\alias{pbernoulli}
\alias{qbernoulli}
\alias{rbernoulli}
\title{Bernoulli distribution wrappers}
\usage{
dbernoulli(n, prob, log = FALSE)

pbernoulli(n, prob, lower.tail = TRUE, log.p = FALSE)

qbernoulli(n, prob, lower.tail = TRUE, log.p = FALSE)

rbernoulli(n, prob)
}
\description{
Wrappers for dbinom, pbinom, qbinom and rbinom. Specify
\code{size = 1}.
}

