% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsim.R
\name{gsim}
\alias{gsim}
\title{Create an initial data.frame for simulation}
\usage{
gsim(data = NULL, target_rows = 1000, seed = 12345)
}
\arguments{
\item{data}{(optional) base dataset. default is NULL in which case a new data.frame will be created.}

\item{target_rows}{(optional) . default is set to 1000 rows}

\item{seed}{(optional) default is 12345. Initial seed}
}
\value{
If data is NULL then a new data.frame is created with sequential index. If data is smaller than target_rows,
       the dataset is expanded with \code{expand_dataset} behaviour. If data is larger than random subset is choosen
       to match target_rows
}
\description{
The gsim() function creates a data.frame with target_rows based on an existing dataset
or a net_new solution without an existing dataset
}
\examples{
gsim(data = mtcars , target_rows = 20)
gsim(target_rows = 2000)
}
