% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as-sims-bayes.R, R/sims-arrays.R
\name{as_sims_array.mcarray}
\alias{as_sims_array}
\alias{as_sims_array.mcarray}
\alias{data_dims}
\alias{is_sims_array}
\alias{is_sims_major}
\alias{n_sims.sims_array}
\alias{sims_array}
\alias{sims_margin}
\title{Construct a simulations array object}
\usage{
\method{as_sims_array}{mcarray}(x, ...)

sims_array(array, sims_major)

is_sims_array(x)

as_sims_array(x, ...)

\method{n_sims}{sims_array}(x)

data_dims(x)

is_sims_major(x)

sims_margin(x)
}
\arguments{
\item{array}{An array of simulations.}

\item{sims_major}{If \code{TRUE}, this indicates that the array is
stored in simulations-major order as opposed to data-major order.}
}
\description{
Construct a simulations array object
}
\seealso{
\code{\link{permute_sims}()}
}

